Optimize Mobile Profile - PR Fixes
Fix the github link in the profile bio. Remove user login name and replace with location – seemed redundant. Remove total fiddles summary – would not display total fiddles, either something in the model is off or my ejs conditional was broken. Either way the information seemed not super valuable so it was nixed.